Switched network to SimpleNetworkWrapper
    Clean commit history version

    Tested briefly, atlas seems to be working (didn't try markers)

    NOTE: Will need to test packets with heavier data loads

    Fixed various generic tags <> missing type arguments
    (these gave me an error when left empty)
