+++ src/main/java/hunternif/mc/atlas/CommonProxy.java
@@ -4,10 +4,12 @@ import hunternif.mc.atlas.ext.ExtTileConfig;
 
 import java.io.File;
 
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.network.simpleimpl.MessageContext;
 
 public class CommonProxy {
 	protected File configDir;
@@ -33,4 +35,11 @@ public class CommonProxy {
 	public void updateBiomeTextureConfig() {}
 	public void updateMarkerTextureConfig() {}
 	public void openAtlasGUI(ItemStack stack) {}
+
+	/**
+	 * Returns a side-appropriate EntityPlayer for use during message handling
+	 */
+	public EntityPlayer getPlayerEntity(MessageContext ctx) {
+		return ctx.getServerHandler().playerEntity;
+	}
 }
