+++ src/main/java/hunternif/mc/atlas/client/SubTileQuartet.java
@@ -42,6 +42,6 @@ public class SubTileQuartet implements Iterable<SubTile> {
 
 	@Override
 	public Iterator<SubTile> iterator() {
-		return new ArrayIterator<>(array);
+		return new ArrayIterator<SubTile>(array);
 	}
 }
